backend/services/ai-service/requirements.txt
# FastAPI dependencies
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.5.0
pydantic-settings>=2.1.0

# LLM Providers
openai>=1.3.0
anthropic>=0.18.0
google-generativeai>=0.3.0

# Database
psycopg2-binary>=2.9.9
sqlalchemy>=2.0.0

# Vector Embeddings
tiktoken>=0.5.0
numpy>=1.24.0

# HTTP Client
httpx>=0.25.0
aiohttp>=3.9.0

# Environment
python-dotenv>=1.0.0

# Redis (for caching and pub/sub)
redis>=5.0.0

# Logging
python-json-logger>=2.0.0

# Celery (for async tasks)
celery>=5.3.0

# Utilities
python-multipart>=0.0.6
aiofiles>=23.2.0

# JWT for authentication
PyJWT>=2.8.0
